Simplify items by using built-in Name component, separate weapons into their own factory, remove EquipmentSlot as standalone component
