+++ src/items/mod.rs
@@ -2,6 +2,8 @@ pub mod components;
 pub mod equipment;
 pub mod inventory;
 pub mod item_factory;
+pub mod weapon_factory;
 
 pub use components::*;
 pub use item_factory::*;
+pub use weapon_factory::*;
