+++ src/ui/pause_menu/button_interactions.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
 use crate::{
-    items::{equipment::Equippable, Consumable, ItemName},
+    items::{equipment::Equippable, Consumable, Item},
     labels::states::PausedState,
     player::Player,
 };
@@ -119,26 +119,18 @@ pub fn handle_inventory_interactions(
 pub fn handle_inventory_click(
     trigger: Trigger<InventoryItemClicked>,
     mut commands: Commands,
-    equipable_query: Query<&Equippable>,
-    consumable_query: Query<&Consumable>,
-    item_name_query: Query<&ItemName>,
+    item_query: Query<(Has<Equippable>, Has<Consumable>), With<Item>>,
     player: Single<Entity, With<Player>>,
 ) {
     if let Some(item_entity) = trigger.item_entity {
-        // Check if item is equipable
-        if let Ok(item_name) = item_name_query.get(item_entity) {
-            debug!("User clicked on {:?}", item_name);
+        if let Ok((equippable, consumable)) = item_query.get(item_entity) {
+            if equippable {
+                commands.trigger_targets(AttemptEquipEvent { item_entity }, *player);
+            } else if consumable {
+                commands.trigger(ConsumeEvent { item_entity });
+            }
         }
 
-        if equipable_query.contains(item_entity) {
-            commands.trigger_targets(AttemptEquipEvent { item_entity }, *player);
-        }
-        // Check if item is consumable
-        else if consumable_query.contains(item_entity) {
-            commands.trigger(ConsumeEvent {
-                item_entity: item_entity,
-            });
-        }
         //Redraw inventory
         commands.trigger(InventoryUpdatedEvent);
     }
