+++ src/ui/pause_menu/inventory_menu.rs
@@ -1,4 +1,4 @@
-use crate::items::{inventory::inventory::Inventory, ItemName};
+use crate::items::{inventory::inventory::Inventory, Item};
 use bevy::prelude::*;
 
 use super::button_interactions::InventoryUpdatedEvent;
@@ -14,7 +14,7 @@ pub struct InventoryDisplay;
 
 pub fn spawn_inventory_menu(
     mut commands: Commands,
-    item_query: Query<&ItemName>,
+    item_query: Query<&Name, With<Item>>,
     player_inventory: Query<&Inventory>,
 ) {
     debug!("spawn_inventory_menu called");
@@ -85,7 +85,7 @@ pub fn spawn_inventory_menu(
                         // Display all inventory items
                         for item_entity in inventory.items.values() {
                             let item_name = item_query.get(*item_entity).unwrap();
-                            spawn_inventory_item(slot_parent, item_name.0.clone(), *item_entity);
+                            spawn_inventory_item(slot_parent, item_name.to_string(), *item_entity);
                         }
 
                         // Display empty slots
@@ -194,7 +194,7 @@ pub fn handle_inventory_update(
     _: Trigger<InventoryUpdatedEvent>,
     mut commands: Commands,
     inventory_menu_query: Query<Entity, With<InventoryMenu>>,
-    item_query: Query<&ItemName>,
+    item_query: Query<&Name, With<Item>>,
     player_inventory: Query<&Inventory>,
 ) {
     // Despawn the existing inventory menu
